implement basic account creation (#7)
This will allow someone to create a new account (with no helpful feedback if something's wrong) via the URL `/account/register`. After the form is displayed, it will be inserted into the database by calling the `CreateAccount` function, which returns the new account id. The new profile's information will be directly queried from the database by its account id and displayed to the user.