Improve Database Abstractions (#8)
Now all of the functions have been loaded directly into the database as functions, which will be called instead of sending raw SQL queries across the db connection.

The `FunctionCall` class has been created to provide an abstraction for function calls. This also provides a fluent API, so database calls in `DbEngine` can just be inline function calls, which are a bit nicer to look at for such a common thing, I believe.